% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_inla.R
\name{augment_inla}
\alias{augment_inla}
\title{Run \code{inla} with new data to obtain new predictors and their positions}
\usage{
augment_inla(.result, newdata)
}
\arguments{
\item{.result}{\code{inla} object.}

\item{newdata}{Dataframe of new data.  The column with the response
(outcome, dependent variable) must be filled with \code{NA}s.}
}
\value{
List with the following items
\describe{
 {inla}{updated \code{inla} object.}
 {new_pos}{positions of the newdata.}
}
}
\description{
Run \code{inla} with new data to obtain new predictors and their positions.
}
\details{
Run \code{inla} with new data to allow the use of \code{summary.linear.predictor},
\code{summary.fitted.values}, \code{marginlas.linear.predictor}, etc. Then
put it in a list with the positions of the newdata described just below.
This function is necessary because \code{INLA} always keep the original data,
as a result we constantly use the row position of the new data when extracting
linear predictors, expected predictors (fitted values), etc.
}
