% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newdata_inla.R
\name{create_newdata_inla}
\alias{create_newdata_inla}
\title{Create a new data set using \code{inla$.args$data} and \code{newdata}}
\usage{
create_newdata_inla(.result, newdata)
}
\arguments{
\item{.result}{\code{inla} object.}

\item{newdata}{Dataframe of new data.  The column with the response
(outcome, dependent variable) must be filled with \code{NA}s.}
}
\value{
List with the following items
\describe{
 {data}{augmented \code{inla$.args$data} with \code{newdata}.}
 {newdata_pos}{positions of the newdata.}
}
}
\description{
Create a new data set using \code{inla$.args$data} and \code{newdata}.
}
\details{
Run \code{inla} with new data to allow the use of \code{summary.linear.predictor},
\code{summary.fitted.values}, \code{marginlas.linear.predictor}, etc.
}
