% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eflINLA.R
\docType{package}
\name{eflINLA}
\alias{eflINLA}
\title{eflINLA: A package with functions working with the \code{INLA} package}
\description{
Utilities to use the \code{INLA} package with the \code{posterior} package.
}
\details{
The functions are mainly used to convert \code{inla} object so that they have
a \code{brms} feel, and, in particular, can be used as \code{rvars} and
\code{draws_vars} with the \code{posterior} package.
}
