% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_marginal.R
\name{extract_marginal}
\alias{extract_marginal}
\title{Get all marginals from \code{inla}}
\usage{
extract_marginal(.result, ren = TRUE)
}
\arguments{
\item{.result}{\code{inla} object}

\item{ren}{Rename the variable with \code{rename_brms}}
}
\value{
List / data.frame of marginals
}
\description{
Get all marginals from \code{inla}.
}
\details{
Extract the hyper and fixed marginals from an \code{inla} object and
put them in a data.frame when \code{is.df = TRUE} (default), otherwise
in a list when \code{is.df = FALSE}.
}
