% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_marginal.R
\name{extract_marginals}
\alias{extract_marginals}
\title{Extract all marginals from \code{inla}}
\usage{
extract_marginals(.result, ren = TRUE)
}
\arguments{
\item{.result}{\code{inla} object.}

\item{ren}{Rename the variable with \code{rename_brms}.}
}
\value{
List of marginals
}
\description{
Extract all marginals from \code{inla}.
}
\details{
Extract the hyper and fixed marginals from an \code{inla} object and
put them in a list.
}
