% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_posterior.R
\name{extract_posterior_samples}
\alias{extract_posterior_samples}
\title{Create list from samples obtained with \code{INLA::inla.posterior.sample}}
\usage{
extract_posterior_samples(samples, type = c("post", "fit", "pred"), sel)
}
\arguments{
\item{samples}{Samples from \code{INLA::inla.posterior.sample}.}

\item{type}{Samples type. One of c("post", "fit", "pred") with the following
meaning:
\describe{
 \item{post}{Get posterior samples, i.e. for variables.}
 \item{fit}{Get posterior samples with predisctors.}
 \item{pred}{Get posterior samples with predisctors and variability.}
}}

\item{sel}{List of selections from \code{posterior_samples_sel}}
}
\value{
List of vector. Length of the list is the number of samples.
}
\description{
Create list from samples obtained with \code{INLA::inla.posterior.sample}.
}
\details{
Create list where the latent and hyper parameters are joined together
in one vector per sample.  Then the n vectors from the n samples are
put together in a list with n items.
}
\seealso{
INLA::inla.posterior.sample eflINLA::posterior_samples_sel
}
