% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Example function that shows message that begins with "Hello" to console.}
\usage{
hello(msg = "Hello, world!")
}
\arguments{
\item{msg}{String with a message that begins with "Hello". Default message
is "Hello, world!".}
}
\value{
A message to console that starts with "Hello"
}
\description{
Example function that shows message that begins with "Hello" to console.
}
