% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_posterior.R
\name{linpred_draws_inla}
\alias{linpred_draws_inla}
\title{Get predictors from samples obtained with \code{INLA::inla.posterior.sample}}
\usage{
linpred_draws_inla(
  .result,
  n = 1L,
  ren = TRUE,
  sel = list(APredictor = 0L, Predictor = 0L),
  pred = FALSE
)
}
\arguments{
\item{.result}{\code{inla} object.}

\item{n}{Sample size.}

\item{ren}{TRUE:rename variables to \code{brms} naming convention;
FALSE: Keep \code{INLA} naming convention.}

\item{sel}{List used for selection by
\code{INLA::inla.posterior.sample(selection = sel)}.}

\item{pred}{TRUE: Add variability to the predictors to make them predictions;
FALSE: Keep predictors as is.}
}
\value{
\code{draws_rvars} object.
}
\description{
Get predictors from samples obtained with \code{INLA::inla.posterior.sample}.
}
\details{
Create list where the predictors are joined together
in one vector per sample.  Then each vector from the \code{n} samples are
put together in a \code{draws_rvars} object.

When \code{pred = FALSE} this is the equivalent of \code{linpred_draws}.
When \code{pred = TRUE} this is the equivalent of \code{predicted_draws}.
}
\seealso{
INLA::inla.posterior.sample
}
