% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_marginal.R
\name{marginal_draws_inla}
\alias{marginal_draws_inla}
\title{Get all marginals from \code{inla} and simulate a sample}
\usage{
marginal_draws_inla(
  .result,
  n = nrow(.result$marginals.fixed[[1]]),
  ren = TRUE,
  repl = TRUE
)
}
\arguments{
\item{.result}{\code{inla} object}

\item{n}{Sample size.}

\item{ren}{Rename the variable with \code{rename_brms}}

\item{repl}{TRUE: Sampling with replacement; FALSE: no replacement.}
}
\value{
List / data.frame of marginals
}
\description{
Get all marginals from \code{inla} and simulate a sample.
}
\details{
Extract the hyper and fixed marginals from an \code{inla} object and
put simulate a sample of size n.
}
