% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{posterior_marg_summary_inla_one}
\alias{posterior_marg_summary_inla_one}
\title{Create one line summary for a given marginal}
\usage{
posterior_marg_summary_inla_one(marg, probs = c(0.025, 0.975))
}
\arguments{
\item{marg}{Marginal from an \code{inla} object.}

\item{probs}{Percentiles returned by the summary.}
}
\value{
vector of summary data.
}
\description{
Create one line summary for a given marginal.
}
\details{
Create one line summary using the marginal funcitons from \code{INLA} such
as \code{inla.emarginal, inla.qmarginal, inla.mmarginal}.
}
