% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{posterior_marg_summmary_inla}
\alias{posterior_marg_summmary_inla}
\title{Write a summary of posterior distribution using marginals}
\usage{
posterior_marg_summmary_inla(.result, probs = c(0.025, 0.975), ren = TRUE)
}
\arguments{
\item{.result}{\code{inla} object.}

\item{probs}{Percentiles returned by the summary.}

\item{ren}{Rename the variable with \code{rename_brms}.}
}
\value{
Dataframe with summaries.
}
\description{
Write a summary of posterior distribution using marginals.
}
\details{
Write a summary of posterior distributions with a format similar to
\code{brms::posterior_summary}.
}
