% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_samples.R
\name{posterior_samples}
\alias{posterior_samples}
\title{Sample the posterior using \code{inla.posterior.sample}}
\usage{
posterior_samples(.result, is.df = TRUE, var = "var")
}
\arguments{
\item{.result}{\code{inla} object.}

\item{is.df}{TRUE: return a data.frame, FALSE: return a list.}

\item{var}{Name of extra column with name of the variable in the data.frame.}
}
\value{
List / data.frame of samples
}
\description{
Sample the posterior using \code{inla.posterior.sample}.
}
\details{
For each sample given by \code{inla.posterior.sample} we extract the value
of the intercept, parameters, precision and hyperparameters. When the
parameters is found in the hyperparameters instead of the \emph{latent} section
the hyperparameters is chosen.
}
