% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_posterior.R
\name{posterior_samples_sel}
\alias{posterior_samples_sel}
\title{Create list to use with \code{INLA::inla.posterior.sample(selection = sel)}}
\usage{
posterior_samples_sel(.result, excl = c("APredictor", "Predictor"))
}
\arguments{
\item{.result}{\code{inla} object.}

\item{excl}{vector of tags to exclude from the selection list. Represents
the predictors. Default is \code{c("APredictor", "Predictor")}.}
}
\value{
List of selected tags and indices
}
\description{
Create list to use with \code{INLA::inla.posterior.sample(selection = sel)}.
}
\details{
The parameters could be in the \emph{hyperpar} section of the sample
as well as in the \emph{latent} section.  This happens when then the
parameter is in the f() section of the inla formula. When this is the case,
this function remove form the selection the variables in \emph{hyperpar}.

See the documentation \code{INLA::inla.posterior.sample} for more details
on the parameters \code{selection}.
}
\seealso{
INLA::inla.posterior.sample
}
