% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_samples.R
\name{posterior_samples_tags}
\alias{posterior_samples_tags}
\title{Find which parameter is in the hyperpar section of the sample}
\usage{
posterior_samples_tags(.result)
}
\arguments{
\item{.result}{\code{inla} object.}
}
\value{
Vector of logical values indicating if the tag is in hyper
}
\description{
Find which parameter is in the hyperpar section of the sample.
}
\details{
The parameters could be in the \emph{hyperpar} section of the sample
as well as in the \emph{latent} section.  When this si the case, we take the
value in \emph{hyperpar}.
}
