% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_01-A.R
\name{posterior_summary}
\alias{posterior_summary}
\title{Get the posterior summary from inla}
\usage{
posterior_summary(r)
}
\arguments{
\item{r}{inla object}
}
\value{
dataframe with inla posterior summary
}
\description{
Get the posterior summary from inla.
}
\details{
Get the posterior summary from inla with precision converted to sd and
quantiles requested by the user.
}
