% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{posterior_summary_inla}
\alias{posterior_summary_inla}
\title{Write a summary of posterior distribution using marginals}
\usage{
posterior_summary_inla(.result, probs = c(0.025, 0.975), ren = TRUE)
}
\arguments{
\item{.result}{\code{inla} object.}

\item{probs}{Percentiles returned by the summary.}

\item{ren}{Rename the variable with \code{rename_brms}.}
}
\value{
Dataframe with summaries.
}
\description{
Write a summary of posterior distribution using marginals.
}
\details{
Write a summary of posterior distributions with a format similar to
\code{brms::posterior_summary}.
}
\examples{
df <- data.frame(
 x1 = rnorm(50, mean = 1, sd = 0.5),
 x2 = rnorm(50, mean = 2, sd = 1),
 sigma = rexp(50, rate = 1))
df$mu <- df$x1 + df$x2
df$y <- rnorm(nrow(df), mean = df$mu, sd = df$sigma)
the_fit <- INLA::inla(formula = y ~ x1 + x2, data = df)
posterior_summary_inla(the_fit)
}
