% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{posterior_summary_inla_one}
\alias{posterior_summary_inla_one}
\title{Create one line summary for a given marginal}
\usage{
posterior_summary_inla_one(marg, probs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{marg}{Marginal from an \code{inla} object.}

\item{probs}{Percentiles returned by the summary. Default is
\code{c(0.025, 0.5, 0.975)} which is the \code{inla} default.}
}
\value{
vector of summary data.
}
\description{
Create one line summary for a given marginal.
}
\details{
Create one line summary using the marginal functions from \code{INLA} such
as \code{inla.emarginal, inla.qmarginal, inla.mmarginal}.
}
\examples{
df <- data.frame(
 x1 = rnorm(50, mean = 1, sd = 0.5),
 x2 = rnorm(50, mean = 2, sd = 1),
 sigma = rexp(50, rate = 1))
df$mu <- df$x1 + df$x2
df$y <- rnorm(nrow(df), mean = df$mu, sd = df$sigma)
the_fit <- INLA::inla(formula = y ~ x1 + x2, data = df)
posterior_summary_inla_one(the_fit$marginals.fixed[[1]])
}
