% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_01-A.R
\name{prec2sd}
\alias{prec2sd}
\title{Convert precision to standard deviation}
\source{
section 2.6, p. 31, Bayesian Inference with INLA, Gomez-Rubio
}
\usage{
prec2sd(marg, is_log = FALSE)
}
\arguments{
\item{marg}{marginal from \code{inla}}

\item{is_log}{FALSE (default): convert using natural scale; TRUE: convert using
log scale.}
}
\value{
precision converted to standard deviation using marginal
}
\description{
Convert precision to standard deviation.
}
\details{
Convert precision to standard deviation with data from the marginal of
INLA.
}
