% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_posterior.R
\name{predict_inla}
\alias{predict_inla}
\title{Get predictors from samples obtained with \code{INLA::inla.posterior.sample}}
\usage{
predict_inla(.result, new_pos = integer(), n = 1L, sel = NULL)
}
\arguments{
\item{.result}{\code{inla} object.}

\item{new_pos}{Integer of positions of \code{newdata}.}

\item{n}{Sample size.}

\item{sel}{List used for selection by
\code{INLA::inla.posterior.sample(selection = sel)}.}
}
\value{
\code{draws_rvars} object.
}
\description{
Get predictors from samples obtained with \code{INLA::inla.posterior.sample}.
}
\details{
Create list where the predictors are joined together
in one vector per sample.  Then each vector from the \code{n} samples are
put together in a \code{draws_rvars} object.
}
\seealso{
INLA::inla.posterior.sample
}
