% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_01-A.R
\name{predict_inla_extract}
\alias{predict_inla_extract}
\title{Extract the data from `inla` posterior samples}
\usage{
predict_inla_extract(samples, pos)
}
\arguments{
\item{samples}{List of samples returned by `inla.posterior.sample`}

\item{pos}{Integer() with the position of the row with dependent variable
set to `NA` in the new data.}
}
\value{
List with 2 lists: the predictors and the sigmas.
}
\description{
Extract the data from `inla` posterior samples to simulate.
}
\details{
Extract the data necessary to add variability to the linear predictors
by simulation.  This is the function that does most of the essential work
in the predict process.
}
