% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_01-A.R
\name{predict_inla_sim}
\alias{predict_inla_sim}
\title{Simulate the predictions from `inla`}
\usage{
predict_inla_sim(sim_data)
}
\arguments{
\item{sim_data}{List with predictors and sigmas. Generated by
`predict_inla_extract`.}
}
\value{
Dataframe of simulations.
}
\description{
Simulate the predictions from `inla`.
}
\details{
Add variability to the predictors by by simulation.
}
