% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_posterior.R
\name{predicted_draws_inla}
\alias{predicted_draws_inla}
\title{Get predictors from samples obtained with \code{INLA::inla.posterior.sample}}
\usage{
predicted_draws_inla(.result, new_pos = integer(), n = 1L, sel = NULL)
}
\arguments{
\item{.result}{\code{inla} object.}

\item{new_pos}{Integer of positions of \code{newdata}.}

\item{n}{Sample size.}

\item{sel}{List used for selection by
\code{INLA::inla.posterior.sample(selection = sel)}.}
}
\value{
\code{draws_rvars} object.
}
\description{
Get predictors from samples obtained with \code{INLA::inla.posterior.sample}.
}
\details{
Extract posterior predictors (as opposed to the ones from marginals) from
an inla object, add variability to them using the standard deviation shown
in each sample \code{INLA::inla.posterior.sample} of size \code{n} and
output these randomized predictors as \code{posterior::draws_vars} objects.

When creating the input inla object, use \code{eflINLA::augment_inla} to
ensure the data is formatted the right way and inla arguments are properly
set.
}
\seealso{
INLA::inla.posterior.sample eflINLA::augment_inla
}
