% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_01-A.R
\name{rename_inla}
\alias{rename_inla}
\title{Rename inla variables}
\usage{
rename_inla(x, choice = c(NA_character_, "Precision"))
}
\arguments{
\item{x}{Character}

\item{choice}{Integer, choice of renaming to perform.}
}
\value{
Modified character
}
\description{
Rename inla variables.
}
\details{
Rename inla variables such as \emph{Precision }to \emph{SD}.
}
\examples{
x <- "Precision for the Gaussian observations"
y <- rename_inla(x, choice = "Precision")
stopifnot(identical(y, "SD for the Gaussian observations"))
}
