% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_01-A.R
\name{rename_inla2brms}
\alias{rename_inla2brms}
\title{Rename strings (row names) from \code{inla} to \code{brms} equivalent}
\usage{
rename_inla2brms(x)
}
\arguments{
\item{x}{Character with \code{inla} strings}
}
\value{
Character with inla strings renamed
}
\description{
Rename strings (row names) from \code{inla} to \code{brms} equivalent.
}
