% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_posterior.R
\name{tidy_draws_inla}
\alias{tidy_draws_inla}
\title{Get a sample of posterior draws from a, \code{inla} fit as a tibble}
\usage{
tidy_draws_inla(.result, n = 1L, ren = TRUE)
}
\arguments{
\item{.result}{\code{inla} object.}

\item{n}{Sample size.}

\item{ren}{TRUE:rename variables to \code{brms} naming convention;
FALSE: Keep \code{INLA} naming convention.}
}
\value{
\code{draws_rvars} object.
}
\description{
Get a sample of posterior draws from a, \code{inla} fit as a tibble. Using
the data from \code{INLA::inla.posterior.sample}.
}
\details{
Create list where the latent and hyper parameters from
\code{INLA::inla.posterior.sample} are joined together in one vector per
sample.  Then each vector from the \code{n} samples are
put together in a \code{draws_rvars} object.
}
\seealso{
INLA::inla.posterior.sample eflINLA::posterior_samples_sel
}
