% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_marginal.R
\name{tidy_marg_draws_inla}
\alias{tidy_marg_draws_inla}
\title{Get fixed and hyper marginals from \code{inla}}
\usage{
tidy_marg_draws_inla(
  .result,
  n = nrow(.result$marginals.fixed[[1]]),
  ren = TRUE,
  repl = TRUE
)
}
\arguments{
\item{.result}{\code{inla} object.}

\item{n}{Sample size.}

\item{ren}{Rename the variable with \code{rename_brms}.}

\item{repl}{TRUE: Sampling with replacement; FALSE: no replacement.}
}
\value{
\code{draws_rvars} object.
}
\description{
Get fixed and hyper marginals from \code{inla}.
}
\details{
Extract the hyper and fixed marginals from an \code{inla} object and
bind them in a \code{draws_rvars} object. The fixed parameter marginals
are obtained from \code{marginals.fixed}. The hyperparameters are obtained
by \code{marginals.hyperpar}.
}
