% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_01-A.R
\name{transform_hyper_marginal}
\alias{transform_hyper_marginal}
\title{Transform the hyperparameters' marginal}
\usage{
transform_hyper_marginal(r, sd_name = "sd")
}
\arguments{
\item{r}{inla object.}

\item{sd_name}{String. Name of standard deviation. Don't change it unless
you have a good reason.}
}
\value{
List of marginals
}
\description{
Transform the hyperparameters' marginal.
}
\details{
Convert the precision used in the hyperparameers' marginals to standard
deviations.
}
