% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_01-A.R
\name{write_hyper_summary}
\alias{write_hyper_summary}
\title{Write summary for inla hyperparameters}
\usage{
write_hyper_summary(margs, qtl = c(0.025, 0.5, 0.975))
}
\arguments{
\item{margs}{Hyperparameters' marginals created with `transform_hyper_marginal`}

\item{qtl}{Numeric(). The quantiles appearing in the summary.}
}
\value{
Dtaframe with hyperparameters' posterior summary
}
\description{
Write summary for inla hyperparameters.
}
\details{
Write a summary with the mean, sd and quantiles of the hyperparameters
from inla. It must use the transformed marginals from `transform_hyper_marginal`
because it requires the standard deviation rather than the precision.  The
format copycat the format using by the inla summaries.
}
\seealso{
transform_hyper_marginal
}
