% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{calc_moments_inla}
\alias{calc_moments_inla}
\title{Compute the moments of a marginal}
\usage{
calc_moments_inla(marg)
}
\arguments{
\item{marg}{Marginal from an \code{inla} object.}
}
\value{
Vector with mean and standard deviation of the marginal.
}
\description{
Compute the moments of a marginal.
}
\details{
Compute the moments of a marginal from \code{inla} using its densities .
}
