% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eflINLA.R
\docType{package}
\name{eflINLA}
\alias{eflINLA}
\title{eflINLA: A package with functions working with the INLA package.}
\description{
The eflINLA package provides functions mainly to extract the data from
the different INLA objects and process them.
}
