% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_marginal.R
\name{linpred_marg_draws_inla}
\alias{linpred_marg_draws_inla}
\title{Draw linear predictors from \code{marginals.linear.predictor}}
\usage{
linpred_marg_draws_inla(
  .result,
  new_pos = integer(),
  n = nrow(.result$marginals.fixed[[1]]),
  ren = TRUE,
  repl = TRUE
)
}
\arguments{
\item{.result}{\code{inla} object.}

\item{new_pos}{Integer of positions of \code{newdata}.}

\item{n}{Sample size (per predictor).}

\item{ren}{TRUE: rename using \code{brms} naming convention.}

\item{repl}{TRUE: Sampling with replacement.}
}
\value{
\code{posterior::draws_rvars} object where every variable is a
linear predictor.
}
\description{
Draw linear predictors from \code{marginals.linear.predictor}.
}
\details{
Draw linear predictors using marginals.
}
