% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_marginals.R
\name{posterior_marginals}
\alias{posterior_marginals}
\title{Get all marginals from \code{inla}}
\usage{
posterior_marginals(.result, is.df = TRUE, var = "var")
}
\arguments{
\item{.result}{\code{inla} object.}

\item{is.df}{TRUE: return a data.frame, FALSE: return a list.}

\item{var}{Name of extra column with name of the variable in the data.frame.}
}
\value{
List / data.frame of marginals
}
\description{
Get all marginals from \code{inla}.
}
\details{
Extract the hyper and fixed marginals from an \code{inla} object and
put them in a data.frame when \code{is.df = TRUE} (default), otherwise
in a list when \code{is.df = FALSE}.
}
