% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_samples.R
\name{posterior_samples}
\alias{posterior_samples}
\title{Sample the posterior using \code{inla.posterior.sample}}
\usage{
posterior_samples(
  .result,
  n = 1L,
  type = c("post", "fit", "pred"),
  out_df = TRUE,
  var = "var",
  ren = TRUE
)
}
\arguments{
\item{.result}{\code{inla} object.}

\item{n}{Sample size.}

\item{type}{Samples type. One of c("post", "fit", "pred") with the following
meaning:
\describe{
 \item{post}{Get posterior samples, i.e. for variables.}
 \item{fit}{Get posterior samples with predisctors.}
 \item{pred}{Get posterior samples with predisctors and variability.}
}}

\item{out_df}{TRUE: return a data.frame, FALSE: return a list.}

\item{var}{Name of extra column with name of the variable in the data.frame.}

\item{ren}{Rename the variable with \code{rename_brms}}
}
\value{
List / data.frame of samples
}
\description{
Sample the posterior using \code{inla.posterior.sample}.
}
\details{
For each sample given by \code{inla.posterior.sample} we extract the value
of the intercept, parameters, precision and hyperparameters. When the
parameters is found in the hyperparameters instead of the \emph{latent} section
the hyperparameters is chosen.
}
\seealso{
INLA::inla.posterior.sample
}
