% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_01-A.R
\name{prec2sd}
\alias{prec2sd}
\title{Convert precision to standard deviation}
\usage{
prec2sd(x, is_log = FALSE)
}
\arguments{
\item{x}{Numeric}

\item{is_log}{FALSE (default): convert using natural scale; TRUE: convert using
log scale.}
}
\value{
precision converted to standard deviation
}
\description{
Convert precision to standard deviation.
}
\details{
Convert precision to standard deviation and inverse trasnform the log when
required.
}
