% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_01-A.R
\name{predict_inla}
\alias{predict_inla}
\title{Predictions using an `inla` model object}
\source{
Bayesian Inference with INLA, Virgilio Gomez-Rubio, 2020, section 2.7,
p. 35-38
}
\usage{
predict_inla(object, pos, n = 1L)
}
\arguments{
\item{object}{Object of class `inla` with the new data.}

\item{pos}{Integer() with the position of the row with dependent variable
set to `NA` in the new data.}

\item{n}{sample size.}
}
\value{
Dataframe with simulated predictions.
}
\description{
Predictions using an `inla` model object with new data.
}
\details{
Make predictions using the `INLA::inla.posterior.sample` function, then
extract the linear predictors and add variability to them.
}
