% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_marginal.R
\name{transform_marginal_hyper}
\alias{transform_marginal_hyper}
\title{Transform the hyperparameters' marginal}
\usage{
transform_marginal_hyper(.result)
}
\arguments{
\item{.result}{inla object.}
}
\value{
List of hyperparameter marginals.
}
\description{
Transform the hyperparameters' marginal.
}
\details{
Convert the precision used in the hyperparameers' marginals to standard
deviations.
}
