% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_marginals.R
\name{transform_marginal_hyper}
\alias{transform_marginal_hyper}
\title{Transform the hyperparameters' marginal}
\usage{
transform_marginal_hyper(.result, sd_name = "sd")
}
\arguments{
\item{.result}{inla object.}

\item{sd_name}{Name of standard deviation. Don't change it unless
you have a good reason.}
}
\value{
List of hyperparameter marginals transformed to user scale
}
\description{
Transform the hyperparameters' marginal.
}
\details{
Convert the precision used in the hyperparameers' marginals to standard
deviations.
}
